[swift] Translate malloc
Summary:
We can map malloc to the internal builtin similar to the free case.

For the arguments of the builtin in llair, we have the following example

```
  int* p = (int*)malloc(sizeof(int));
```

translated to

```
 %2!5 := builtin malloc(4);	; test/analyze/free_free.c:11:18:
```

where I belive that 4 is an int expression, so we use the typ part of that expression, so we build a Textual.Exp.Typ expression that the translation to sil expects.

Reviewed By: ngorogiannis

Differential Revision: D70187658

fbshipit-source-id: 374be4f4cecb06c4d2c1e9506374a6d6ff1a7065