[swift] Llair to Textual: translate FuncName
Summary: There isn't a way in Textual to keep the information that this is a function. But if this is used in a call with the Indirect call target, we are adding the expression as the first argument of the builtin "llvm_dynamic_call", so we can process it there.

Reviewed By: davidpichardie

Differential Revision: D70257048

fbshipit-source-id: 4bab3e5507fa8daab6665e42a860f5d82d5bc15b