[topl] Better filtering
Summary:
If function f calls function g and both have what appears to be the same
issue, Topl reports only on g. The "same issue" is detected by a
heuristic. Previously, the issues had to have the same pulse-latentness
and topl-latentness. The idea being that if g has a latent issue that
becomes manifest in f, then you want to report both: they are not the
"same issue".

However, the following situation arises in practice:
  g has a manifest issue
  f has a latent issue, because it calls g only sometimes
In this case, it is better if the latent issue is hidden.

In this diff, we make manifest issues (like the one in g) hide latent
issues (like the one in f).

Reviewed By: hajduakos

Differential Revision: D70256788

fbshipit-source-id: d4b5e3fdcd681ee12481cb62994a9056c9484f33