[inferpython][capture] simplifing the handling of POP_TOP bytecode
Summary:
this bytecode should pop the top of the stack. I don't understand Vincent's code here.
If the computation that returned this value was side-effectful, it should have generated an instruction before. This is not the job of this step to do that. We apply this methodology on `Yield` that indeed has side-effects.

Note: I'm not sure what to do with YieldFrom for now. We do not have analysis tests about it anyway. And its bytecode representation seems to change in Python 3.12 so no rush.

Reviewed By: dulmarod

Differential Revision:
D69658503

Privacy Context Container: L1208441

fbshipit-source-id: 9fbdeae34241d89302f2ed6d051f6d8d803e6cf1