[swift] Llair to Textual: translate Global and Reg
Summary:
We still need to find a way to figure out when to use the id and when to use the name. In Llair Reg is used everywhere but in textual we sometimes need Var and sometimes LVar. We need examples to figure out a solution, just translating it as Var for now.

For globals, I'm assuming just translating them like this, and them being added to the list of globals should be enough.

Reviewed By: ngorogiannis

Differential Revision: D70330863

fbshipit-source-id: 67212b55b3532715e97182e6cd1d8d4b34dec49b