[InferPython] model compare_in for dicts
Summary:
We want to avoid false positives like:
```
d = {"key1": 1, "key2": 2}
if "key3" in d:
  key3 = d["key3"]
```

For that we provide a model for py_compare_in which checks first the type of the first argument. If is a constant string, we check if the second argument is a dict and try to find the corresponding key. In any other situation the operator behaves as before (unknown).

Reviewed By: davidpichardie

Differential Revision:
D70567971

Privacy Context Container: L1208441

fbshipit-source-id: 70b8c1c3db113cb74c67375601634cc3eb009c22