[dbwriter] use file locking instead of separate process
Summary:
Infer currently uses a separate process to do all the database writes, so as to avoid the problems of multiple concurrent writers that SQLite has. Every worker that needs to do a write connects to the database writer, serialises its payload, sends it to the writer, which deserialises it and synchronously does the write.

We can avoid paying the cost of serialisation, sending and deserialisation by having workers do the writes themselves, ensuring mutual exclusion. We implement this by using `lockf` on a lock file stored in the results directory.

Tests show a 20% reduction of analysis wall time, and this gain seems to be uniform over the number of workers.

Reviewed By: martintrojer

Differential Revision:
D70903084

Privacy Context Container: L1208441

fbshipit-source-id: 6bebc138ece1106a7a879288b5d4125379dc322b