[inferpython][OO] instance attributes inherited from class companion
Summary: Here we focus on attribute looking. First, we use dynamic type to recognize if we have a class instance. Then, if we are sure the attribute is not in the instance, we search into its companion class. We do not deal with inheritance yet.

Reviewed By: Gabisampaio

Differential Revision:
D71192646

Privacy Context Container: L1208441

fbshipit-source-id: edec823c2add94cdd0348ae5d6855d304a9d90cd