[llvm] Improving translating of variables
Summary:
The locals in llair are all the variables used, which does not correspond to the locals we use in Infer. Trying to identify the locals, as variables that have been "allocated".  Also, we translate a variable as a `Var` if it isn't a local or a formal, and as a Lvar otherwise.

- We skip `alloc`, just using it as a marker to call a variable local.
- If the store tries to store a local or formal variable, do a load beforehand.

Reviewed By: ngorogiannis

Differential Revision: D71486273

fbshipit-source-id: f04bdb2dc9a83986ec4e73220e13c1d503fb3de3