[Autofixes] Add new fields for autofixes to jsonbug.atd
Summary:
We will soon add more involved autofixes for Hack so it makes sense to change the current shape used.
To avoid breaking the current pipeline for autofixes, for now we keep both old and new shapes. We will start populating the new `autofixes` and `autofix_candidates` fields and then delete the previous `autofix` one.

Now we consider each autofix to be a list of transformations (not making the distinction of `additional` that existed before). Each transformation contains a file, line, column, original and replacement.

There are then two types of autofixes: approved autofixes (named as `autofixes`) and candidates (`autofix_candidates`), each being a list of autofixes. The approved autofixes are used currently for obj-c where no additional filtering is required. The candidates will be used for Hack where we need to make sure the changes typecheck afterwards and potentially filter out bad candidates.

Reviewed By: martintrojer

Differential Revision:
D71466920

Privacy Context Container: L1208441

fbshipit-source-id: a554a36b9edeb34f250a365a0edc44921ddda012