[multicore] optionally enable global checker state
Summary: Checkers register global state save and restore functions when ondemand goes in and out of a procedure, even when these checkers are not running. This is a bit wasteful, and in the case of biabduction which has not been updated to work with multicore, this leads to data races. This diff makes the registration contingent on the checker being enabled.

Reviewed By: davidpichardie

Differential Revision: D71696831

fbshipit-source-id: 0868160ecdb114274de4d6f8b5ea288cdbe8a177