[inferpython] improve location positionning
Summary: while developping the new assertino-error checker I observed some bad locations on nodes and instructions. We improve the PyIR transformation by propagating the exit location of the previous node into the next node to build. It will serves as default location until we find a better location (if any in the node).

Reviewed By: ngorogiannis

Differential Revision: D71616793

fbshipit-source-id: 3aa281801d16f7aad8533d1a27e5b7f7bbcda674