[llvm] Do not translate free
Summary: In llair we have the free expression after both `alloc` and `malloc`, but for infer we only need it after `malloc`. So we need to adjust this on the llair side. Removing all the free functinos for now so we can continue translating some examples.

Reviewed By: ngorogiannis

Differential Revision:
D71636009

Privacy Context Container: L1208441

fbshipit-source-id: 46e694d1e3bfa9f3fbccbcaa87652e9b1e385a44