[llvm] Keep track of types of known ids
Summary:
To continue doing type inference we also need to keep track of the types of ids. In particular, we want to update the type of ids that may be opaque pointers with a pointer of the correct type, based on the type present in the load instruction.

A further pass of type inference is needed after to update the code of previous instructions, and previous variables.

Reviewed By: ngorogiannis

Differential Revision:
D71739907

Privacy Context Container: L1208441

fbshipit-source-id: e748fc31a136898aa44dbb202be091dc694bc624