[inferpython] new AssertionError checker
Summary:
for now we will primilary use that for writing in a simpler way some of our unit tests, but
this could be a cool checker in our portfolio anyway.

I decided to not make the signal latent. See explanations in the diff, and in comments in the tests.

Reviewed By: ngorogiannis

Differential Revision:
D71539706

Privacy Context Container: L1208441

fbshipit-source-id: dbf5bbf54670684a2b8f15e2b777412bbc72166d