[topl] Log error for summaries with "free" vars
Summary:
A Topl prune condition of a summary should only mention abstract values
(vars) that appear in registers or in the corresponding Pulse summary.
Such summaries lead to latent issues that *cannot* become manifest.
Essentially, it means Topl wastes time looking at problems that cannot
happen, making it more likely that it misses problems that can happen.

Let's log when such bad summaries exist, so we can fix it.

The easiest is to log at the point where such a summary is applied.

Reviewed By: hajduakos

Differential Revision: D70972675

fbshipit-source-id: c22867eff596830ea0b8efafee562663c1b45317