[topl] Simplify the check for what Pulse constraints
Summary:
Topl is supposed to drop simple-states whose registers refer to values
that are not known in any way to Pulse. Previously the check for "known
to Pulse" was rather complicated, including reachability that possibly
took into account attributes, etc. Now it simply checks if the
(abstract) value syntactically occurs in pre, post, or path condition.

Also, removed some code that was unused for a long time.

Reviewed By: hajduakos

Differential Revision: D71129576

fbshipit-source-id: a87a152fea99976f668808aa95dbed88ee53e185