[topl] Drop ambiguous simple states
Summary:
The main goal of this diff is to ensure that Topl summaries do not
mention "local" variables, which the callee cannot influence. This
change gets rid of a few false positives in the tests. It also
introduces a false negative in the sil/topl tests. I suspect that kind
of situation occurs rarely. (After all, it is SIL test precisely because
I couldn't figure out how to write a test in a high-level language that
would exercise the prover in a certain way.)

In this commit message, I say that a variable is *local* when it cannot
be unified with a caller variable. This commit assumes that the
non-local variables are (a) those mentioned in the pre Pulse state and
(b) those mentioned in the pre Topl registers; all others are assumed local.

We drop Topl simple states that prune on local variables in the
`filter_for_summary` function, which is called at the end of turning
(abstract) states into summaries.

Reviewed By: hajduakos

Differential Revision: D71205816

fbshipit-source-id: 4778f7ac423c7a9357628d1cc36056eaef33dec5