[llvm] Improve location of statements with default location
Summary: There are some statements that do not belong to code in the source code. In our examples this corresponds to loading the formals into local variables. In that case, those statements get the location [0, 0]. So now we keep around a procedure state, which will probably be useful in general, and add the procedure location. Then we use that location instead of [0, 0]. That will be more useful in case of a report in those lines.

Reviewed By: davidpichardie

Differential Revision:
D71910518

Privacy Context Container: L1208441

fbshipit-source-id: c0b6b3cd008635a1bf6683c983b58d703af2e387