Optionally dedup issues by trace
Summary:
`Errlog` has a generic deduplication mechanism, which did not allow two
issues of the same type and with the same description to be reported at
the same location. This is problematic for example when Topl is used to
find a flow. Such flows are attached to the location of the sink, and
there is no variation in the description of the issue for different
sources. But traces are distinct.

In general, I think dedup should be based on the information presented
to the user: most users would not complain to see two distinct traces
reported at the same location; but they would probably complain if they
are shown two seemingly identical issues.

Still, I left the old behavior as the default. We can switch the defaul
later.

Reviewed By: ngorogiannis

Differential Revision: D71816536

fbshipit-source-id: 197a16c0c0153a8233c1f06045f46ca4e469469c