[topl] Add annotation pattern
Summary:
The pattern for procedure calls now includes an (optional) annotation
pattern. If an annotation pattern is provided it only matches Java
procedures that have an annotation matching the given regexp.

```
stateA -> stateB: @ "AnnotRegexp" "ProcedureRegexp" (ArgA, ArgB, Ret) ...
```
Like all other regexps in Topl, the annotation regexp can be negated by
preceding it with `!`. The annotation pattern itself can also be
negated, by preceding it with `~`. Here is a tricky example:
```
A -> B: ~@!".*foo.*" procname(Ret)
```
This pattern matches for procedures called `procname` that do not have
any annotation that contains the substring "foo"; that is, it matches
when all the annotations on `procname` contain the substring "foo".

The annotations being considered are: the annotation on the return value
of `procname`, the annotations on the class in which `procname` is
defined, and the annotations on this class's base classes.

Reviewed By: hajduakos

Differential Revision: D71903399

fbshipit-source-id: 2b26083c49d4a4cb358622245c4d05831e749235