[llvm] Type inference for variables
Summary:
At the end of the translation of a node, use the types inferred for variables to update all the variables with better types.

Textual already has type inference implemented, so it's easier to focus on getting the types of the variables right, and not add types to instructions, which Textual then can do with the correct types for variables.

Works on Load and Store for now. We'll look into other instructions later.

Reviewed By: ngorogiannis

Differential Revision:
D71988066

Privacy Context Container: L1208441

fbshipit-source-id: d20bf3e34526ce0221658eb1d7e041ea39dc10d1