[llvm] More tests about function calls and conditions
Summary: Further tests that translate correctly, we also get the expected NPEs in Pulse. There was a small mistake in the location of the node, we were checking that the original instructions were empty, whilst we should check that the translated instructions are, as the lists are not always the same length.

Reviewed By: ngorogiannis

Differential Revision: D72176042

fbshipit-source-id: c2b4976217710f19cdfd5ad8180b5afcbd8b09b5