[InferPython] change error loc of uaa
Summary:
Some UAA reports are a bit confusing as we report the location where the resource becomes unreachable. The line content of this location is sometimes just a closing parenthesis and it becomes difficult for the user to debug properly.
With this diff users should see the error on the allocation line instead, which may help understanding what the error is about.

Reviewed By: davidpichardie

Differential Revision:
D72313689

Privacy Context Container: L1208441

fbshipit-source-id: eb37f724ccc8ea60fa1f90370bb8a687d759e3ac