[topl] Add automaton info to traces
Summary:
The old traces pointed only to the program. The description of each step
was essentially a short version of what could already be seen in the
program source. (So the utility of these descriptions was minor: to
notice if somehow the traces appear to point to the wrong location in
the code, for example because of errors in computing line numbers due to
macro expansion.)

Now the descriptions contain more information. In particular, they
describe the effect a program statement has on the user provided
property.

Reviewed By: hajduakos

Differential Revision: D72395413

fbshipit-source-id: 5ac65dcfc2edecd9c9879c31ce5406e527d70624