[inferpython][type inference][debug] adding unit tests to show type inference
Summary:
the current type inference engine is badly engineered. It should not be performed at
Textual level but at PyIR level where we are more guided by Ocaml type system. In this diff, I add debug in order to observe the refactoring that will be performed in the later diff.
The final goal is to remove all the inference type from PyIR2Textuel.ml

Reviewed By: ngorogiannis

Differential Revision:
D71965511

Privacy Context Container: L1208441

fbshipit-source-id: 8c0c7c4d0271ea66aa5a49e511a718f5c1c4daba