[topl] Treat nonnull predicates as soft
Summary:
What to do if a user-defined constrain in not proven? Ther are a few
options: (a) delay so that we gain more info from a caller contex (i.e.,
make it latent), or (b) hide the issue. Prior to this change, Topl chose
between (a) and (b) based on whether it thought that gaining more info
from a caller context could possibly help.

This change splits case (b) into two subcases: (b1) the constraint is
'soft' which mean we report the issue, and (b2) the constraint is 'hard'
which mean we hide the issue. Intuitively, 'hard' constraints must be
proven, while 'soft' constraints only prune cases where their negation
can be proven.

So what is 'soft' and what is 'hard'? For now, we use a simple
heuristic: if the --topl-nonnull-soft is given then constraints of the
form `X!=0` are 'soft'; otherwise all constraints are 'hard' (as it was
before). This is similar to how Pulse has some special treatment for
these kinds of constraints when deciding whether an issue is latent or
manifest.

In the future, we might expose this soft/hard distinction in the
property language.

Reviewed By: hajduakos

Differential Revision: D72645298

fbshipit-source-id: 4a8c72d0529fed1a310d0caa2c97cc0173b40a8a