[llvm] Adding original names to variables
Summary:
We add the original name as the name of variables when they are available.
This just takes directly what is in llvm, which is that the names for parameters are given to a local copy instead. We may be able to transform this local copy to eliminate it and give the correct name to the parameters instead.

Reviewed By: ngorogiannis

Differential Revision:
D72636652

Privacy Context Container: L1208441

fbshipit-source-id: c7475c0d6a1ce971bab7b0ee35c0ee2e50012805