[stdcompat] upgrade and stop vendoring
Summary:
Stdcompat has been updated to be compatible with OCaml 5.2 so we don't need to vendor and pin the package anymore.

NB we don't have to specify the stdcompat version in infer.opam since the only one compatible with OCaml 5.2 is the one we need.

Reviewed By: jvillard

Differential Revision:
D72705466

Privacy Context Container: L1208441

fbshipit-source-id: 60594d8f5f302bc65399f7e90918b023328d308a