[llvm] Translation of structs
Summary:
Now that we have a type environment in llair, we translate it to Textual.
We also add the struct_map to the proc_state and pass it around  to differentiate when we need to translate the struct type vs. when we it's in the map already.
This map may also be useful in the future to get the field names in the field access expressions.

Reviewed By: ngorogiannis

Differential Revision:
D72865530

Privacy Context Container: L1208441

fbshipit-source-id: 0225224a37f51ac88c63decbe460e227599e2e0e