[topl] On missing files, don't crash
Summary:
If the user provides a bad property, we want to not crash but only print
a warning. If the user runs in capture mode, for example, topl
properties are not really needed, so it makes sense to chug along.

The implementation is a bit complicated by the fact that Config does not
have access to Logging (which is done to avoid a module cycle).

Reviewed By: hajduakos

Differential Revision: D72792223

fbshipit-source-id: a1f537a1d9fa187471064c2db161224f6f456ef3