[inferll] un-vendor llvm bindings and upgrade to llvm-18
Summary:
Vendoring llvm-dune is cumbersome because it makes it hard to keep track
of the patches we have on top of it. On the other hand it's convenient
for us to change llvm bindings as we wish. The latter is now less
convenient but the former is much easier and made upgrading to llvm-18 a
breeze.

I couldn't make llvm-17 as in opam work due to weird errors from
opam/dune/linkers conspiring against my sanity, so we get the upgrade to
llvm-18 in this commit too instead.

Reviewed By: dulmarod

Differential Revision: D72864615

fbshipit-source-id: afbce48bd33f1f27e1882daff0d115a8848521b0