[llvm] use a new local opam repo to set up llvm for inferll
Summary:
This makes it a bit easier to maintain: the setup-llvm.sh script is now
just about adding the opam repo and asking opam to install the package.
It will also be easier to upstream our changes, eg to add a new package
for llvm20, and easier to have several packages in there (eg
conf-llvm.20 since it's not in opam either yet).

Reviewed By: dulmarod

Differential Revision:
D73112359

Privacy Context Container: L1208441

fbshipit-source-id: 2557df183ca2dced50a1598769acf7fdfd654f80