[llvm] upgrade to llvm20 and use the LLVM distribution in facebook-clang-plugins
Summary:
llvm-20 is the version that we use for the clang frontend. Let's use the
same one for the LLVM frontend so as to install only one version of LLVM
at a time.

Moved to the -static version of opam llvm because I couldn't make the
-shared one build: `llvm-config` seemingly still has this bug where it
doesn't find the shared libs in libLLVM.so and instead insists of trying
to find individual modules' .so files. Oh well.

Most of the build stuff from llvm18 is reused as-is on llvm20 and just
works, so that's nice.

Reviewed By: ngorogiannis

Differential Revision:
D73112357

Privacy Context Container: L1208441

fbshipit-source-id: 164a40f65d9b086bf221d6688b0b049474201f4f