[java][pulse] set dynamic type of parameters with a final type
Summary:
When a parameter as a declared type A with A a final class, we know for sure its dynamic type is A.
This information will enable more virtual call resolution, when combined with summary specialization. See tests.

Thanks Raquel Fernandes da Silva for your bug report!

Reviewed By: ngorogiannis

Differential Revision:
D73111566

Privacy Context Container: L1208441

fbshipit-source-id: bb038745582d40937e867d13ff8fc18d64cd9dc5