[cpp] request dyntype specialization at every shallow copy
Summary:
when a `std::function<...>`function is copied, the current Cpp models use a Pulse shallow_copy operation to make the assignment. But if at that time the closure value is unknown (meaning we don't know it's dynamic_type yet), the copy is very poorly informative. It happens for example when you assign fields in a constructor. Later, when the summary of the constructor is called (for example), you may now now the dynamic_type of the source, but this is too late. The fix I used here is: when copying a closure, if the dyntype is not known yet, ask for it using specialization. This approach fix the problem but it will generate specialization of the constructor for each closure creation point. This is expensive.

Something to try
1) change all the copy models around closures
2) or, improve the Pulse solver to reasons on symbolic equality like typeof(X)==typeof(Y)

Reviewed By: ngorogiannis

Differential Revision:
D73183749

Privacy Context Container: L1208441

fbshipit-source-id: 0687351c595b3f9009744c49e444e1a10aba87de