[llvm] Fix the name of fields in structs
Summary: Before we were using the offset that llair uses which corresponds to number of bytes. In Infer, we don't want to do pointer arithmetic, we want to have just the position of the field as the identifier. Later on we'll try to get  the real names from the debug info.

Reviewed By: martintrojer

Differential Revision: D74008706

fbshipit-source-id: d43ad14b7324ec93608395c2dd4c985190909bae