[multicore] reduce mutex contention on lazy values
Summary:
The current implementation of thread safe evaluation introduces unnecessary contention between threads: for already-forced values, threads still take turns on a mutex for read-only access.

This diff uses an atomic reference and a union type that represents either an already forced value, or an unforced one that is paired with a mutex. The mutex is used (plus double-checked locking) when forcing the value, after which the atomic reference is updated to point to an  unadorned plain value, and the mutex is GCed.

Reviewed By: martintrojer

Differential Revision:
D74992069

Privacy Context Container: L1208441

fbshipit-source-id: cdb9aca399b59404d3befee52da87ac04cfa7402