[swift] Add Swift language to Textual
Summary:
We can now use the Swift procname in the translation to Sil, and assign the Swift language to the sil files.

The Swift procname uses Mangled names which are built with a plain and a mangled name. There is always a mangled name, but there may not be a plain name, in that case we use only the mangled name.

Reviewed By: ngorogiannis

Differential Revision: D75007279

fbshipit-source-id: 7d9b00ba41de1245bfb651b3dfb3c0dcee03059b