[Rust] Rust support (#1905)
Summary:
# Rust Support
This PR adds the initial setup for integrating rust in infer. Currently it calls a rust program that outputs the rust Intermediate Representation (MIR) of a file to analyze, eventually this MIR will be translated to textual for analysis.
The goal will be to detect memory errors in unsafe rust.

Examples of memory errors in rust:
```rust
let ptr: *const i32 = std::ptr::null();
unsafe {
    let x : i32 = *ptr; //Null dereference here
}
```
```rust
let ptr: *const i32;
unsafe {
    {
        let x = Box::new(50); // Store 50 on the heap
        ptr = &*x;
    }
    // x dropped as it goes out of scope, rust will clean up x and free its memory on the heap
    let y = *ptr; // Undefined behavior: use after free
}
```
# Technical
In infer/lib/rust is the rust program that will translate a rust program to textual.
It requires access to the internal rust compiler crate which are available on nightly builds of rust.
To run this program rustup can be installed https://rustup.rs/ to install the rust language and the rust package manager cargo.
The program is called from infer/src/integration/Rust.ml

# Next Steps
- Defining textual definitions in lib/rust/src/textual
- Translations from MIR to textual.
- Adding test procedures

Pull Request resolved: https://github.com/facebook/infer/pull/1905

Reviewed By: ngorogiannis

Differential Revision: D76038831

Pulled By: davidpichardie

fbshipit-source-id: 826f68a1d14c8f0088392e2e5a795bbebaf86bbe