[llvm] Improve the creadion of ids
Summary: The current creation of ids relied on the ids of llair's registers which had two problems: it was not local to the method, and we didn't have a good way of creating ids if there wasn't a matching llair register, which we'll need in the next diff. This solution solves that.

Reviewed By: davidpichardie

Differential Revision: D76048232

fbshipit-source-id: cf50c41803589f85c49d6ad52f2dade1ba0c6746