[llvm] Add llair translation for llvm.expect and improve translation of true
Summary:
We want to translate assertions. They are translated using the builtin `llvm.expect`, some jumps and the assert function which is called in case of a failure. Here we translate the builtin `llvm.expect` and also improve the translation of `true`, which was `-1` before.

Still need more improvements to handle assert: using the model, and ignore the real code.

Reviewed By: davidpichardie

Differential Revision:
D76276624

Privacy Context Container: L1208441

fbshipit-source-id: 8b1ccf96f9b8f7f7959010c8f35fc4f83ac04a1f