[llvm] Only translate code in the given source file
Summary: For this to work the paths need to be complete, so we needed to add the full path to the Makefiles for the tests. Also the file and dir values in a llair loc are different in Swift and C, so we need to add a separate case for each language.

Reviewed By: ngorogiannis

Differential Revision: D76351147

fbshipit-source-id: 4d15716685284cec709a32c8202f911620708998