[swift] First steps in translating classes
Summary:
Some cast assertions fail for Swift, removing those. This fails in the given test that includes classes that we'll start translating.
To avoid this small example crashing we also needed to improve the Procname.pp to be actually verbose. Morevoer, we do not need to translate the code of deinit (destructors) functions.

Reviewed By: ngorogiannis

Differential Revision: D76509760

fbshipit-source-id: 26ec077010a5ae770d3bab7400fa5383f8aa56e1