[swift] Remove calls to swift_release
Summary:
We won't be tracking retain counts so let's remove the calls to retain/release to simplify the code. Only release in this diff, retain can come later when it appears in some example.
We do this directly in sledge translation because we want to avoid creating terminators for no reason. Then, on the infer side, we remove calls to nondet when there is no return variable as this is redundant.
Also added an option to print types in the proc_state as the type env is getting too big to print in some places.

Reviewed By: davidpichardie

Differential Revision: D76736727

fbshipit-source-id: cea27c7ef9b1372432b98388ac16767709473bdc