[swift] Improving adding dereference instructions
Summary: Adding Load instructions to load fields more consistently. Translating the types of classes as pointers to structs.  This is how we do things generally in Infer in other languages. Also means adding a Load instruction when a field is accessed with `extractvalue` which gets translated to `Move`. This gets rid of most typechecking errors in the examples. One left to be fixed in the next diff.

Reviewed By: davidpichardie

Differential Revision: D77021121

fbshipit-source-id: 1f0818cf18535484ebe0b5a1577ef13cfc3d170f