[swift] Improve the special case of removing the instruction for storing zero into an object
Summary: We still want to remove this store of zero in a class, but now the class is represented as a pointer to struct, which, because of opaque pointers, could be a pointer to int. So we need to take care of that case too. Also, to get that type, we can extract it from the type environment.

Reviewed By: davidpichardie

Differential Revision: D77022898

fbshipit-source-id: d9b3c57b70547a0c62bd1b0c72352995c9588cb5