[swift] Translation of records
Summary: This translates record as a list of instructions for setting each field of a variable, together with the variable. For this to work in globals however, we need to translate the init of the globals in its own initialiser method. To be done in a follow-up diff as this is a separate problem to the actual translation of records.

Reviewed By: hajduakos

Differential Revision: D77799733

fbshipit-source-id: 3a4a2e631047421e4f3c6709da12a1bc89e80499