[infinite recursion] fixing arguments equality test in cycle
Summary:
when a mutual recursion cycle is detected (ie. a function `foo` appears twice in a callchain), we decide to send an INFINITE_RECURSION signal by comparing the arguments between the fst and the snd occurence of `foo` in the callchain.
```
function foo(x, y) {
   bar(y,x,0)
}

function bar(u, v, w) {
  foo(v, u)
}
```

This comparison was not properly done currently since we were not using the right abstract values for arguments. Now we properly use the `actuals` fields we have in PulseMutualRecursion module.

We also remove the complex logic Jules introduced around test around `has_reachable_in_inner_pre_heap`. This was introduced before INFINITE_RECUSION (when only MUTUAL_RECURSION_CYCLE existed) and is useless now.

Reviewed By: Gabisampaio

Differential Revision: D79161302

fbshipit-source-id: f53e90c749d098132227cbb12ef8a18f8ea312e9