[llvm] Refactoring to be able to do builtin calls as expressions
Summary: `to_textual_call_aux` was returning only instructions, which makes it difficult to use when we want to make an expression be translated as a call. This refactoring allows to call this functions from `to_textual_exp` to translate expressions as calls (builtins in this case). To be done in the next diff.

Reviewed By: ngorogiannis

Differential Revision: D79889687

fbshipit-source-id: 4fd4eca88629c26606ec2a56912a46bd69ad7d35