[llvm] Translate GetElementPtr as a builtin
Summary: The llvm has a internal global with metadata about classes, which includes the method names of the class in a particular index. Given that we may not know the type of an object at compile time, specially because of opaque ptr types in llvm, we will translate GetElementPtr as a builtin, and will try to find the correct method to call at runtime when we have the dynamic type of the object. We will need to store the index to method correspondence in the tenv.

Reviewed By: ngorogiannis

Differential Revision: D79894475

fbshipit-source-id: 8e9f0358eef320bf3f6bcdb1f71945de1fe4537c