[devai_quality] Python comparison excluding type annotations (v0)
Summary: This can be useful to check if two programs are equal. Only type annotations are excluded for now, which means any other change (even a renaming which doesn't change program behaviour) will be flagged as different. In the future we may need to tune this to exclude some behaviour preserving changes.

Reviewed By: davidpichardie

Differential Revision:
D80269745

Privacy Context Container: L1208441

fbshipit-source-id: 9942f5b8280f7d88a91fd79eecbd326da1dcb8ef