[semdiff] ignore import statements
Summary: it can happen in python that type annotations are added together with import statements. We want to avoid raising an issue on those cases, so for now we ignore import statements. We can have a false negative scenario (see test) if the import statement itself changes program behaviour, but this might be less common.

Reviewed By: davidpichardie

Differential Revision:
D80793817

Privacy Context Container: L1208441

fbshipit-source-id: 101c8acddbb24b5ac7141a0e1d52bc6bbe315170