[semdiff] avoid calling Py.initialize at top level
Summary: Py.initialize() can't be called twice and this was creating a problem when adding a dependency in next diff (D81061264). We change to avoid calling Py.initialize as soon as OCaml library is loaded. Instead we invoke the function only when `semdiff` is actually used. We also check if it has been called before already to avoid errors.

Reviewed By: ngorogiannis

Differential Revision:
D81218066

Privacy Context Container: L1208441

fbshipit-source-id: 0a260d5944493779207de3386342e47dadf8e1cb