[textual][conditionnal expr][4/N] refactoring around fresh_ident
Summary:
some Textual transformation need to add new fresh ident. We save, when
possible, a fresh ident in each procdesc, and make sure it is uptodate

For now, there is only one transformation which need this information, but we will add a new
one in a next diff.

Reviewed By: ngorogiannis

Differential Revision:
D80611030

Privacy Context Container: L1208441

fbshipit-source-id: 44ef3bbb271a6a9c019c94ecbccb2d698c97ce72