[swift] Fix typing error
Summary: See last diff for context on the problem. We add a load in the case of `Var id` too but only when it's a pointer. The issue is to find a reliable way of getting the type. Some llvm types are wrong. So now we only add the inferred types to  the map of ids (only in the translation of `move` for now) and use those types in `add_deref`.

Reviewed By: davidpichardie

Differential Revision: D81485473

fbshipit-source-id: 374a3f0f8a5b82b79ea7dcea5eeedc81dca5f80b