[timer] reimplement using Mtime
Summary: The timer module used Unix.gettimes to get the current time, which leads to contention around a futex in multicore mode when all threads are racing on it. We change the implementation to use the existing Mtime module, which provides access to the clock_gettime system call which doesn't seem to have the same drawbacks. We also re-enable timeouts on multicore as a result.

Reviewed By: davidpichardie

Differential Revision:
D81702853

Privacy Context Container: L1208441

fbshipit-source-id: 84bdf3d002bdd3c06bbe8a38baf5c01d78caaa36