[hash] stop using primes as initial capacity
Summary:
In the codebase, there are many instances where hashtables are created with a prime number as the initial capacity. The standard library implementation does not benefit from this and it, in fact, rounds up to the nearest power of two.

https://github.com/lucasaiu/ocaml/blob/323378a510bda2abd463cfa1888deb9a965e9e87/stdlib/hashtbl.ml#L60

To avoid spreading this myth by way of preexisting code, we change primes to the smallest power of 2 applicable.

Reviewed By: dulmarod

Differential Revision:
D81780890

Privacy Context Container: L1208441

fbshipit-source-id: bff69340aae085068b187a86bcb6d153c2756a99