[swift] Turn off translating global initialisers for now
Summary: There are many internal globals that we don't need. Moreover, even for the ones we need, translating their initialisers properly into Sil is not currently a priority. Separating the internal globals out is not obvious currently because the bindings for accessing attributes of globals are not there. So I'm removing this translation for all globals for now, but keeping the mechanism there for when this becomes a priority. It was made to compute the offset of methods, but this has now been achieved.

Reviewed By: davidpichardie

Differential Revision: D81783390

fbshipit-source-id: ae2d11debc9d80507a84f276f2ab38e72d74a175