[semdiff] add replacer for class/isinstance
Summary: sometimes diffs differ in the sense that some "safe transformations" happened, but behaviour could still considered to be preserved. One example is replacing __class__ by insinstance() operator in python. We handle this case now and in the future we may come up with a more generic approach for this problem.

Reviewed By: davidpichardie

Differential Revision:
D82106211

Privacy Context Container: L1208441

fbshipit-source-id: 5aaf0dff01bef2800184c8ac3be4948879daa7bf