[semdiff] change output diff format
Summary: Showing the ast diff as before is not much useful because it's something hard for the human to read. We now change the output to be like a "git diff". We first compare the ASTs and get the lines which differ and then show the corresponding lines of the program.

Reviewed By: davidpichardie

Differential Revision:
D82302868

Privacy Context Container: L1208441

fbshipit-source-id: 5f8f6aef2772af9c6c3cc96b61e17ee5fc24b742